Move EncodingTable and CodePageDataItem to System.Text namespace
This is the more logical namespace for it and it matches the namespace that these live under in CoreRT.
