+++ src/mscorlib/src/System/Text/CodePageDataItem.Unix.cs
@@ -2,7 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-namespace System.Globalization
+namespace System.Text
 {
     internal class CodePageDataItem
     {
